🏗️ Architect: Modularized apts/events.py
Refactored the 1,200+ line 'God Object' apts/events.py into a modular package structure:
- Split logic into rarity.py, translator.py, and calculations/ sub-package (lunar, planetary, space, sky).
- Refactored AstronomicalEvents in coordinator.py to delegate to these modules.
- Optimized get_events with a dispatch dictionary to reduce cyclomatic complexity.
- Maintained backward compatibility for public API and internal methods used by tests.
- Updated all unit test mocks to reflect the new structure.
- Verified with full test suite (631 passed).
