--- apts/events.py
@@ -1,1258 +1,0 @@
-import logging
-import time
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from datetime import datetime, timezone
-from itertools import combinations
-from typing import List, Optional
-
-import numpy as np
-import pandas as pd
-import requests
-import requests_cache
-from dateutil.parser import parse as parse_date
-from skyfield import almanac
-from skyfield.api import Star, Topos
-
-from . import cache, skyfield_searches
-from .cache import get_ephemeris, get_timescale
-from .catalogs import Catalogs
-from .config import get_event_settings
-from .constants.event_types import EventType
-from .i18n import get_language, gettext_
-from .utils import planetary
-
-logger = logging.getLogger(__name__)
-
-requests_cache.install_cache("apts_cache", backend="memory", expire_after=300)
-
-utc = timezone.utc
-
-
-class AstronomicalEvents:
-    # Optimization: Only pre-compute positions if any conjunction-related events are enabled
-    # to avoid significant overhead when only basic events (like Moon phases) are requested.
-    CONJUNCTION_EVENTS = [
-        "moon_messier_conjunctions",
-        "moon_star_conjunctions",
-        "planet_messier_conjunctions",
-        "planet_star_conjunctions",
-        "planetary_dichotomy",
-    ]
-
-    def __init__(
-        self,
-        place,
-        start_date,
-        end_date,
-        events_to_calculate: Optional[List[EventType]] = None,
-    ):
-        # Oracle Persona Boundary: restrict predictions to +/- 100 years.
-        current_year = datetime.now(timezone.utc).year
-        if (
-            abs(start_date.year - current_year) > 100
-            or abs(end_date.year - current_year) > 100
-        ):
-            raise ValueError(
-                f"Oracle: Prediction range restricted to +/- 100 years from current date ({current_year})."
-            )
-
-        self.place = place
-        self.start_date = start_date.astimezone(utc)  # Ensure start_date is UTC
-        self.end_date = end_date.astimezone(utc)  # Ensure end_date is UTC
-        self.ts = get_timescale()
-        self.eph = get_ephemeris()
-        self.observer = self.eph["earth"] + Topos(
-            latitude_degrees=self.place.lat_decimal,
-            longitude_degrees=self.place.lon_decimal,
-            elevation_m=self.place.elevation,
-        )
-        self.events = []
-        # Load global settings from config
-        config_settings = get_event_settings()
-
-        # Default settings: most are True, some expensive ones are False
-        self.event_settings = {event.value: True for event in EventType}
-        # Disable expensive events by default
-        self.event_settings["golden_hour"] = False
-        self.event_settings["blue_hour"] = False
-        self.event_settings["culminations"] = False
-        self.event_settings["jovian_moon_events"] = False
-
-        # Override with config settings if present
-        for key, value in config_settings.items():
-            if key in self.event_settings:
-                self.event_settings[key] = value
-
-        logger.debug(
-            f"AstronomicalEvents initialized. Config-based settings: {self.event_settings}"
-        )
-
-        if events_to_calculate is not None:
-            # If a list of events is provided, it should explicitly enable those
-            # but ONLY if they are not explicitly disabled in config.
-            # This allows tests to force-enable events (as they have no config)
-            # while respecting user settings in real environments.
-            events_to_calculate_str = [str(e) for e in events_to_calculate]
-            logger.debug(f"Limiting to events_to_calculate: {events_to_calculate_str}")
-            new_settings = {event.value: False for event in EventType}
-            for e_str in events_to_calculate_str:
-                # Use config if present, otherwise default to True for the requested event
-                new_settings[e_str] = config_settings.get(e_str, True)
-            self.event_settings = new_settings
-            logger.debug(f"Final event_settings after override: {self.event_settings}")
-
-        self.executor = ThreadPoolExecutor()
-        self.catalogs = Catalogs()
-
-    def shutdown(self):
-        self.executor.shutdown(wait=True)
-
-    def get_events(self):
-        start_time = time.time()
-        executor = self.executor
-        futures = []
-
-        # Pre-compute positions for moving bodies often used in multiple searches
-        # to improve overall calculation efficiency.
-        precomputed = {}
-        if any(self.event_settings.get(e) for e in self.CONJUNCTION_EVENTS):
-            num_hours = int((self.end_date - self.start_date).total_seconds() / 3600)
-            if num_hours >= 2:
-                times = self.ts.linspace(
-                    self.ts.utc(self.start_date), self.ts.utc(self.end_date), num_hours
-                )
-                moon_obj = planetary.get_skyfield_obj("moon")
-                # Using .apparent() to share high-precision positions
-                # Optimization: Hoist observer.at(times) out of the loop
-                obs_at_times = self.observer.at(times)
-                precomputed["moon"] = obs_at_times.observe(moon_obj).apparent()
-
-                # Pre-compute positions for major planets
-                for p_name in planetary.CONJUNCTION_PLANETS:
-                    p_obj = planetary.get_skyfield_obj(p_name)
-                    precomputed[p_name.lower()] = obs_at_times.observe(p_obj).apparent()
-
-        if self.event_settings.get("moon_phases"):
-            futures.append(executor.submit(self.calculate_moon_phases))
-        if self.event_settings.get("conjunctions"):
-            futures.append(executor.submit(self.calculate_conjunctions))
-        if self.event_settings.get("oppositions"):
-            futures.append(executor.submit(self.calculate_oppositions))
-        if self.event_settings.get("meteor_showers"):
-            futures.append(executor.submit(self.calculate_meteor_showers))
-        if self.event_settings.get("highest_altitudes"):
-            futures.append(executor.submit(self.calculate_highest_altitudes))
-        if self.event_settings.get("lunar_occultations"):
-            futures.append(executor.submit(self.calculate_lunar_occultations))
-        if self.event_settings.get("aphelion_perihelion"):
-            futures.append(executor.submit(self.calculate_aphelion_perihelion))
-        if self.event_settings.get("moon_apogee_perigee"):
-            futures.append(executor.submit(self.calculate_moon_apogee_perigee))
-        if self.event_settings.get("mercury_inferior_conjunctions"):
-            futures.append(
-                executor.submit(self.calculate_mercury_inferior_conjunctions)
-            )
-        if self.event_settings.get("moon_messier_conjunctions"):
-            futures.append(
-                executor.submit(self.calculate_moon_messier_conjunctions, precomputed)
-            )
-        if self.event_settings.get("moon_star_conjunctions"):
-            futures.append(
-                executor.submit(self.calculate_moon_star_conjunctions, precomputed)
-            )
-        if self.event_settings.get("space_launches"):
-            futures.append(executor.submit(self.calculate_space_launches))
-        if self.event_settings.get("space_events"):
-            futures.append(executor.submit(self.calculate_space_events))
-        if self.event_settings.get("iss_flybys"):
-            futures.append(executor.submit(self.calculate_iss_flybys))
-        if self.event_settings.get("tiangong_flybys"):
-            futures.append(executor.submit(self.calculate_tiangong_flybys))
-        if self.event_settings.get("solar_eclipses"):
-            futures.append(executor.submit(self.calculate_solar_eclipses))
-        if self.event_settings.get("lunar_eclipses"):
-            futures.append(executor.submit(self.calculate_lunar_eclipses))
-        if self.event_settings.get("nasa_comets"):
-            futures.append(executor.submit(self.calculate_nasa_comets))
-        if self.event_settings.get("planet_alignments"):
-            futures.append(executor.submit(self.calculate_planet_alignments))
-        if self.event_settings.get("lunar_planetary_occultations"):
-            futures.append(executor.submit(self.calculate_lunar_planetary_occultations))
-        if self.event_settings.get("messier_culminations"):
-            futures.append(executor.submit(self.calculate_messier_culminations))
-        if self.event_settings.get("jovian_moon_events"):
-            futures.append(executor.submit(self.calculate_jovian_moon_events))
-        if self.event_settings.get("saturn_ring_crossings"):
-            futures.append(executor.submit(self.calculate_saturn_ring_crossings))
-        if self.event_settings.get("jupiter_grs_transits"):
-            futures.append(executor.submit(self.calculate_jupiter_grs_transits))
-        if self.event_settings.get("planet_messier_conjunctions"):
-            futures.append(
-                executor.submit(self.calculate_planet_messier_conjunctions, precomputed)
-            )
-        if self.event_settings.get("planet_star_conjunctions"):
-            futures.append(
-                executor.submit(self.calculate_planet_star_conjunctions, precomputed)
-            )
-        if self.event_settings.get("planet_stationary_points"):
-            futures.append(executor.submit(self.calculate_planet_stationary_points))
-        if self.event_settings.get("planet_solar_conjunctions"):
-            futures.append(executor.submit(self.calculate_planet_solar_conjunctions))
-        if self.event_settings.get("lunar_features"):
-            futures.append(executor.submit(self.calculate_lunar_features))
-        if self.event_settings.get("planet_planet_occultations"):
-            futures.append(executor.submit(self.calculate_planet_planet_occultations))
-        if self.event_settings.get("venus_great_brilliancy"):
-            futures.append(executor.submit(self.calculate_venus_greatest_brilliancy))
-        if self.event_settings.get("supermoons"):
-            futures.append(executor.submit(self.calculate_supermoons))
-        if self.event_settings.get("mars_closest_approach"):
-            futures.append(executor.submit(self.calculate_mars_closest_approach))
-        if self.event_settings.get("jovian_mutual_events"):
-            futures.append(executor.submit(self.calculate_jovian_mutual_events))
-
-        # Golden hour, Blue hour and Culminations are disabled by default as they generate too many events
-        if self.event_settings.get("golden_hour", False) or self.event_settings.get(
-            "blue_hour", False
-        ):
-            futures.append(executor.submit(self.calculate_golden_blue_hours))
-        if self.event_settings.get("culminations", False):
-            futures.append(executor.submit(self.calculate_culminations))
-
-        if self.event_settings.get("greatest_elongations"):
-            futures.append(executor.submit(self.calculate_greatest_elongations))
-        if self.event_settings.get("planetary_dichotomy"):
-            futures.append(
-                executor.submit(self.calculate_planetary_dichotomy, precomputed)
-            )
-        if self.event_settings.get("seasons"):
-            futures.append(executor.submit(self.calculate_seasons))
-
-        for future in as_completed(futures):
-            try:
-                self.events.extend(future.result())
-            except Exception as e:
-                logger.error(f"Error calculating astronomical events: {e}")
-
-        # Standard columns for the events DataFrame
-        columns = [
-            "date",
-            "event",
-            "type",
-            "rarity",
-            "object",
-            "object1",
-            "object2",
-            "altitude",
-            "planets",
-            "separation_degrees",
-            "phase",
-            "shower_name",
-            "distance_km",
-        ]
-
-        if not self.events:
-            return pd.DataFrame(columns=pd.Index(columns))
-
-        df = pd.DataFrame(self.events)
-        # Ensure all standard columns exist
-        for col in columns:
-            if col not in df.columns:
-                df[col] = None
-
-        df = df.sort_values(by="date")
-        df = self.translate_events(df)
-        logger.debug(f"--- get_events: {time.time() - start_time}s")
-        return df
-
-    def translate_events(self, df: pd.DataFrame) -> pd.DataFrame:
-        if df.empty:
-            return df
-
-        # Optimization: Early return if language is English (no translation needed)
-        if get_language() == "en":
-            return df
-
-        columns_to_translate = [
-            "event",
-            "type",
-            "shower_name",
-            "phase",
-            "object",
-            "object1",
-            "object2",
-            "planets",
-        ]
-        for col in columns_to_translate:
-            if col in df.columns:
-                # Optimization: Use unique value mapping to avoid O(N) gettext_ calls.
-                # This provides a massive speedup (up to 15x) for large datasets with repeated strings.
-                # We use a custom unique collection for potential unhashable types (lists).
-                try:
-                    unique_values = df[col].unique()
-                except TypeError:
-                    # Fallback for columns containing unhashable types (e.g., lists)
-                    unique_values = []
-                    seen = set()
-                    for item in df[col]:
-                        h_item = tuple(item) if isinstance(item, list) else item
-                        if h_item not in seen:
-                            seen.add(h_item)
-                            unique_values.append(item)
-
-                translation_map = {}
-                has_list = False
-                for val in unique_values:
-                    if isinstance(val, list):
-                        translation_map[tuple(val)] = [gettext_(item) for item in val]
-                        has_list = True
-                    elif isinstance(val, str):
-                        translation_map[val] = gettext_(val)
-                    else:
-                        translation_map[val] = val
-
-                if has_list:
-                    # Lists are not hashable, so we use tuples for the mapping but still
-                    # need to handle input list types correctly in apply.
-                    df[col] = df[col].apply(
-                        lambda x: (
-                            translation_map[tuple(x)]
-                            if isinstance(x, list)
-                            else translation_map.get(x, x)
-                        )
-                    )
-                else:
-                    df[col] = df[col].apply(lambda x: translation_map.get(x, x))
-
-        return df
-
-    def _get_rarity(self, event_type: str, data: dict) -> int:
-        if event_type == "Moon Phase":
-            return 1
-        if event_type in [
-            "Conjunction",
-            "Moon-Messier Conjunction",
-            "Moon-Star Conjunction",
-            "Planet-Messier Conjunction",
-            "Planet-Star Conjunction",
-        ]:
-            sep = data.get("separation_degrees", 5.0)
-            if sep < 0.15:
-                return 5
-            if sep < 0.4:
-                return 4
-            if sep < 1.2:
-                return 3
-            if sep < 2.5:
-                return 2
-            return 1
-        if event_type == "Opposition":
-            obj = data.get("object", "")
-            if obj in ["Mars", "Jupiter", "Saturn", "Uranus", "Neptune"]:
-                return 4
-            return 3
-        if event_type == "Meteor Shower":
-            if data.get("phase") == "Peak":
-                return 4
-            return 1
-        if event_type == "Planet Altitude":
-            return 3
-        if event_type in ["Lunar Occultation", "Lunar Planetary Occultation"]:
-            return 4
-        if event_type == "Aphelion/Perihelion":
-            return 1
-        if event_type == "Moon Apogee/Perigee":
-            return 1
-        if event_type == "Inferior Conjunction":
-            if data.get("is_transit"):
-                return 5
-            return 3
-        if event_type == "Solar Eclipse":
-            kind = str(data.get("eclipse_type", "")).lower()
-            if "total" in kind:
-                return 5
-            if "annular" in kind:
-                return 5
-            return 4
-        if event_type == "Lunar Eclipse":
-            kind = str(data.get("eclipse_kind", "")).lower()
-            if "total" in kind:
-                return 5
-            if "partial" in kind:
-                return 4
-            return 3
-        if event_type in ["Space Launch", "Space Event"]:
-            return 2
-        if event_type in ["ISS Flyby", "Tiangong Flyby"]:
-            mag = data.get("peak_magnitude", 0)
-            if mag < -3:
-                return 4
-            if mag < -1:
-                return 3
-            return 1
-        if event_type == "Comet":
-            return 5
-        if event_type == "Planet Alignment":
-            planets_count = len(data.get("planets", []))
-            if planets_count >= 6:
-                return 5
-            if planets_count == 5:
-                return 4
-            if planets_count == 4:
-                return 3
-            return 2
-        if event_type == "Greatest Elongation":
-            return 3
-        if event_type == "Season":
-            return 2
-        if event_type == "Messier Culmination":
-            return 2
-        if event_type == "Jovian Moon Event":
-            return 3
-        if event_type == "Jupiter GRS Transit":
-            return 1
-        if event_type == "Planet Stationary Point":
-            return 3
-        if event_type == "Planet Solar Conjunction":
-            return 3
-        if event_type == "Lunar Feature":
-            return 4
-        if event_type == "Planet-Planet Occultation":
-            return 5
-        if event_type == "Venus Greatest Brilliancy":
-            return 4
-        if event_type == "Supermoon":
-            return 3
-        if event_type == "Planetary Dichotomy":
-            return 3
-        if event_type == "Mars Closest Approach":
-            return 4
-        return 1
-
-    def calculate_venus_greatest_brilliancy(self):
-        start_time = time.time()
-        events = skyfield_searches.find_venus_greatest_brilliancy(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Venus Greatest Brilliancy", event)
-        logger.debug(
-            f"--- calculate_venus_greatest_brilliancy: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_space_launches(self):
-        start_time = time.time()
-        events = []
-        start_date_str = self.start_date.strftime("%Y-%m-%d")
-        end_date_str = self.end_date.strftime("%Y-%m-%d")
-        url = f"https://ll.thespacedevs.com/2.2.0/launch/upcoming/?window_start__gte={start_date_str}&window_end__lte={end_date_str}"
-        try:
-            response = requests.get(url, timeout=10)
-            response.raise_for_status()
-            data = response.json()
-            for launch in data.get("results", []):
-                event_data = {
-                    "date": parse_date(launch["window_start"]).astimezone(utc),
-                    "event": launch["name"],
-                    "type": "Space Launch",
-                }
-                event_data["rarity"] = self._get_rarity("Space Launch", event_data)
-                events.append(event_data)
-        except requests.exceptions.RequestException as e:
-            logger.error(f"Error fetching space launches: {e}")
-        logger.debug(f"--- calculate_space_launches: {time.time() - start_time}s")
-        return events
-
-    def calculate_space_events(self):
-        start_time = time.time()
-        events = []
-        start_date_str = self.start_date.strftime("%Y-%m-%d")
-        end_date_str = self.end_date.strftime("%Y-%m-%d")
-        url = f"https://ll.thespacedevs.com/2.2.0/event/upcoming/?date__gte={start_date_str}&date__lte={end_date_str}"
-        try:
-            response = requests.get(url, timeout=10)
-            response.raise_for_status()
-            data = response.json()
-            for event in data.get("results", []):
-                event_data = {
-                    "date": parse_date(event["date"]).astimezone(utc),
-                    "event": event["name"],
-                    "type": "Space Event",
-                }
-                event_data["rarity"] = self._get_rarity("Space Event", event_data)
-                events.append(event_data)
-        except requests.exceptions.RequestException as e:
-            logger.error(f"Error fetching space events: {e}")
-        logger.debug(f"--- calculate_space_events: {time.time() - start_time}s")
-        return events
-
-    def calculate_iss_flybys(self):
-        start_time = time.time()
-        # Create the Topos object directly from place attributes
-        topos_observer = Topos(
-            latitude_degrees=self.place.lat_decimal,
-            longitude_degrees=self.place.lon_decimal,
-            elevation_m=self.place.elevation,
-        )
-        events = skyfield_searches.find_iss_flybys(
-            topos_observer, self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("ISS Flyby", event)
-        logger.debug(f"--- calculate_iss_flybys: {time.time() - start_time}s")
-        return events
-
-    def calculate_tiangong_flybys(self):
-        start_time = time.time()
-        # Create the Topos object directly from place attributes
-        topos_observer = Topos(
-            latitude_degrees=self.place.lat_decimal,
-            longitude_degrees=self.place.lon_decimal,
-            elevation_m=self.place.elevation,
-        )
-        events = skyfield_searches.find_tiangong_flybys(
-            topos_observer, self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Tiangong Flyby", event)
-        logger.debug(f"--- calculate_tiangong_flybys: {time.time() - start_time}s")
-        return events
-
-    def calculate_solar_eclipses(self):
-        start_time = time.time()
-        events = skyfield_searches.find_solar_eclipses(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            kind = event.get("eclipse_type", "")
-            event["event"] = f"{kind} Solar Eclipse"
-            event["type"] = "Solar Eclipse"
-            event["rarity"] = self._get_rarity("Solar Eclipse", event)
-        logger.debug(f"--- calculate_solar_eclipses: {time.time() - start_time}s")
-        return events
-
-    def calculate_lunar_eclipses(self):
-        start_time = time.time()
-        events = skyfield_searches.find_lunar_eclipses(
-            self.start_date, self.end_date, observer=self.observer
-        )
-        for event in events:
-            event["event"] = "Lunar Eclipse"
-            event["type"] = "Lunar Eclipse"
-            event["rarity"] = self._get_rarity("Lunar Eclipse", event)
-        logger.debug(f"--- calculate_lunar_eclipses: {time.time() - start_time}s")
-        return events
-
-    def calculate_moon_phases(self):
-        start_time = time.time()
-        events = []
-        t0 = self.ts.utc(self.start_date)
-        t1 = self.ts.utc(self.end_date)
-        which_phase = almanac.moon_phases(self.eph)
-        t, y = almanac.find_discrete(t0, t1, which_phase)
-        for ti, yi in zip(t, y):
-            event_data = {
-                "date": ti.utc_datetime().astimezone(utc),
-                "event": almanac.MOON_PHASES[yi],
-                "type": "Moon Phase",
-            }
-            event_data["rarity"] = self._get_rarity("Moon Phase", event_data)
-            events.append(event_data)
-        logger.debug(f"--- calculate_moon_phases: {time.time() - start_time}s")
-        return events
-
-    def calculate_conjunctions(self):
-        start_time = time.time()
-        events = []
-        planets_data = {
-            p: planetary.get_skyfield_obj(p) for p in planetary.CONJUNCTION_PLANETS
-        }
-        moon = "moon"
-        moon_display_name = "Moon"
-        moon_obj = planetary.get_skyfield_obj(moon)
-
-        # Pre-compute positions for all bodies involved in conjunctions
-        # Using 0.01 days (~14.4 minutes) to match the Moon's resolution in searches
-        step = 0.01
-        num_steps = int(
-            (self.end_date - self.start_date).total_seconds() / (step * 86400)
-        )
-        if num_steps < 2:
-            num_steps = 2
-        times = self.ts.linspace(
-            self.ts.utc(self.start_date), self.ts.utc(self.end_date), num_steps
-        )
-
-        precomputed = {}
-        # Pre-observe all planets and the moon at the given times
-        # Note: observer.at(times).observe(obj) returns an Astrometric object
-        # that behaves like an array of positions.
-        # Optimization: Hoist observer.at(times) out of the loop
-        obs_at_times = self.observer.at(times)
-        for p, obj in planets_data.items():
-            precomputed[p.lower()] = obs_at_times.observe(obj)
-        precomputed[moon.lower()] = obs_at_times.observe(moon_obj)
-
-        executor = self.executor
-        futures = {}
-
-        # Optimization: Planet-Planet conjunctions
-        # Vectorized search across pairs to avoid overhead of multiple iterative searches
-        for (p1, p1_obj), (p2, p2_obj) in combinations(planets_data.items(), 2):
-            p1_name = planetary.get_simple_name(p1)
-            p2_name = planetary.get_simple_name(p2)
-            future = executor.submit(
-                skyfield_searches.find_conjunctions_between_moving_bodies,
-                self.observer,
-                p1,
-                [(p2_name, p2_obj)],
-                self.start_date,
-                self.end_date,
-                threshold_degrees=5.0,
-                precomputed_positions=precomputed,
-            )
-            futures[future] = p1_name
-
-        # Planet-Moon conjunctions
-        planet_list = [
-            (planetary.get_simple_name(p), obj) for p, obj in planets_data.items()
-        ]
-        future_moon = executor.submit(
-            skyfield_searches.find_conjunctions_between_moving_bodies,
-            self.observer,
-            moon,
-            planet_list,
-            self.start_date,
-            self.end_date,
-            threshold_degrees=5.0,
-            precomputed_positions=precomputed,
-        )
-        futures[future_moon] = moon_display_name
-
-        for future in as_completed(futures):
-            obj1_name = futures[future]
-            found_events = future.result()
-            for event in found_events:
-                event["type"] = "Conjunction"
-                event["event"] = "Conjunction"
-                event["object1"] = obj1_name
-                # object2 is already in the event dict from find_conjunctions_between_moving_bodies
-                event["rarity"] = self._get_rarity("Conjunction", event)
-                events.append(event)
-        logger.debug(f"--- calculate_conjunctions: {time.time() - start_time}s")
-        return events
-
-    def calculate_oppositions(self):
-        start_time = time.time()
-        events = []
-        planets = planetary.OPPOSITION_PLANETS
-        executor = self.executor
-        futures = [
-            executor.submit(
-                skyfield_searches.find_oppositions,
-                self.observer,
-                p,
-                self.start_date,
-                self.end_date,
-            )
-            for p in planets
-        ]
-        for future in as_completed(futures):
-            found_events = future.result()
-            for event in found_events:
-                event["type"] = "Opposition"
-                simple_name = planetary.get_simple_name(event["planet"])
-                event["event"] = "Opposition"
-                event["object"] = simple_name
-                del event["planet"]
-                event["rarity"] = self._get_rarity("Opposition", event)
-            events.extend(found_events)
-        logger.debug(f"--- calculate_oppositions: {time.time() - start_time}s")
-        return events
-
-    def calculate_meteor_showers(self):
-        start_time = time.time()
-        events = skyfield_searches.find_meteor_showers(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            # Ensure dates are timezone-aware UTC
-            event["date"] = event["date"].astimezone(utc)
-            event["rarity"] = self._get_rarity("Meteor Shower", event)
-        logger.debug(f"--- calculate_meteor_showers: {time.time() - start_time}s")
-        return events
-
-    def calculate_highest_altitudes(self):
-        start_time = time.time()
-        events = []
-        executor = self.executor
-        futures = {
-            executor.submit(
-                skyfield_searches.find_highest_altitude,
-                self.observer,
-                planetary.get_skyfield_obj(planet_name),
-                self.start_date,
-                self.end_date,
-            ): planet_name
-            for planet_name in ["mercury", "venus"]
-        }
-        for future in as_completed(futures):
-            planet_name = futures[future]
-            t, alt = future.result()
-            if t:
-                simple_name = planetary.get_simple_name(planet_name)
-                event_data = {
-                    "date": t.astimezone(utc),  # type: ignore
-                    "event": "Highest altitude",
-                    "object": simple_name,
-                    "type": "Planet Altitude",
-                    "altitude": alt,
-                }
-                event_data["rarity"] = self._get_rarity("Planet Altitude", event_data)
-                events.append(event_data)
-        logger.debug(f"--- calculate_highest_altitudes: {time.time() - start_time}s")
-        return events
-
-    def calculate_lunar_occultations(self):
-        start_time = time.time()
-        events = skyfield_searches.find_lunar_occultations(
-            self.observer,
-            self.catalogs.BRIGHT_STARS,
-            self.start_date,
-            self.end_date,
-        )
-        for event in events:
-            event["type"] = "Lunar Occultation"
-            event["event"] = "Lunar Occultation"
-            event["rarity"] = self._get_rarity("Lunar Occultation", event)
-        logger.debug(f"--- calculate_lunar_occultations: {time.time() - start_time}s")
-        return events
-
-    def calculate_lunar_planetary_occultations(self):
-        start_time = time.time()
-        events = skyfield_searches.find_lunar_planetary_occultations(
-            self.observer,
-            self.start_date,
-            self.end_date,
-        )
-        for event in events:
-            event["type"] = "Lunar Planetary Occultation"
-            event["event"] = "Lunar Planetary Occultation"
-            event["rarity"] = self._get_rarity("Lunar Planetary Occultation", event)
-        logger.debug(
-            f"--- calculate_lunar_planetary_occultations: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_aphelion_perihelion(self):
-        start_time = time.time()
-        events = []
-        planets = planetary.APHELION_PERIHELION_PLANETS
-        executor = self.executor
-        futures = [
-            executor.submit(
-                skyfield_searches.find_aphelion_perihelion,
-                planet_name,
-                self.start_date,
-                self.end_date,
-            )
-            for planet_name in planets
-        ]
-        for future in as_completed(futures):
-            events_from_search = future.result()
-            for event_dict in events_from_search:
-                simple_name = planetary.get_simple_name(event_dict["planet"])
-                event_dict["event"] = event_dict["event_type"]
-                event_dict["object"] = simple_name
-                del event_dict["planet"]
-                del event_dict["event_type"]
-                event_dict["date"] = event_dict["date"].astimezone(utc)  # type: ignore
-                event_dict["type"] = "Aphelion/Perihelion"
-                event_dict["rarity"] = self._get_rarity(
-                    "Aphelion/Perihelion", event_dict
-                )
-                events.append(event_dict)
-        logger.debug(f"--- calculate_aphelion_perihelion: {time.time() - start_time}s")
-        return events
-
-    def calculate_moon_apogee_perigee(self):
-        start_time = time.time()
-        events = skyfield_searches.find_moon_apogee_perigee(
-            self.start_date, self.end_date
-        )
-        for event in events:
-            event["type"] = "Moon Apogee/Perigee"
-            event["rarity"] = self._get_rarity("Moon Apogee/Perigee", event)
-        logger.debug(f"--- calculate_moon_apogee_perigee: {time.time() - start_time}s")
-        return events
-
-    def calculate_mercury_inferior_conjunctions(self):
-        start_time = time.time()
-        events = skyfield_searches.find_mercury_inferior_conjunctions(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["type"] = "Inferior Conjunction"
-            if event.get("is_transit"):
-                event["event"] = "Mercury Transit"
-            else:
-                event["event"] = "Inferior Conjunction"
-            event["object"] = "Mercury"
-            event["rarity"] = self._get_rarity("Inferior Conjunction", event)
-        logger.debug(
-            f"--- calculate_mercury_inferior_conjunctions: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_moon_messier_conjunctions(self, precomputed_positions=None):
-        start_time = time.time()
-        messier_objects_to_check = [
-            "M1",
-            "M2",
-            "M3",
-            "M4",
-            "M5",
-            "M8",
-            "M9",
-            "M10",
-            "M11",
-            "M12",
-            "M14",
-            "M15",
-            "M16",
-            "M17",
-            "M18",
-            "M19",
-            "M20",
-            "M21",
-            "M22",
-            "M23",
-            "M24",
-            "M25",
-            "M26",
-            "M27",
-            "M28",
-            "M30",
-            "M35",
-            "M41",
-            "M42",
-            "M43",
-            "M44",
-            "M45",
-            "M46",
-            "M47",
-            "M48",
-            "M49",
-            "M50",
-            "M53",
-            "M54",
-            "M55",
-            "M58",
-            "M59",
-            "M60",
-            "M61",
-            "M62",
-            "M64",
-            "M65",
-            "M66",
-            "M67",
-            "M68",
-            "M71",
-            "M72",
-            "M73",
-            "M74",
-            "M75",
-            "M77",
-            "M78",
-            "M79",
-            "M80",
-            "M83",
-            "M84",
-            "M85",
-            "M86",
-            "M87",
-            "M88",
-            "M89",
-            "M90",
-            "M91",
-            "M93",
-            "M95",
-            "M96",
-            "M98",
-            "M99",
-            "M100",
-            "M104",
-            "M105",
-            "M107",
-        ]
-
-        messier_df = self.catalogs.MESSIER[
-            self.catalogs.MESSIER["Messier"].isin(messier_objects_to_check)
-        ]
-
-        star_data = list(zip(messier_df["Messier"], messier_df["skyfield_object"]))
-
-        conjunctions = skyfield_searches.find_conjunctions_with_stars(
-            self.observer,
-            "moon",
-            star_data,
-            self.start_date,
-            self.end_date,
-            threshold_degrees=4.0,
-            precomputed_positions=precomputed_positions,
-        )
-
-        events = []
-        for conj in conjunctions:
-            event_data = {
-                "date": conj["date"].astimezone(utc),
-                "event": "Conjunction",
-                "object1": "Moon",
-                "object2": conj["object2"],
-                "separation_degrees": conj["separation_degrees"],
-                "type": "Moon-Messier Conjunction",
-            }
-            event_data["rarity"] = self._get_rarity(
-                "Moon-Messier Conjunction", event_data
-            )
-            events.append(event_data)
-
-        logger.debug(
-            f"--- calculate_moon_messier_conjunctions: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_moon_star_conjunctions(self, precomputed_positions=None):
-        start_time = time.time()
-
-        # Filter stars close to the ecliptic (within 10 degrees)
-        # The Moon stays within ~5.1 degrees of the ecliptic.
-        ts = get_timescale()
-        t_ref = ts.utc(self.start_date)
-
-        star_objs_all = self.catalogs.BRIGHT_STARS["skyfield_object"].tolist()
-        star_names_all = self.catalogs.BRIGHT_STARS["Name"].tolist()
-
-        # Observe all stars at once using a vectorized Star object
-        # Optimization: use pre-calculated float columns instead of list comprehension over Star objects
-        stars_vector = Star(
-            ra_hours=self.catalogs.BRIGHT_STARS["ra_hours"].to_numpy(),
-            dec_degrees=self.catalogs.BRIGHT_STARS["dec_degrees"].to_numpy(),
-        )
-        spos_at_t_ref = self.observer.at(t_ref).observe(stars_vector)
-        lats, _, _ = spos_at_t_ref.ecliptic_latlon()
-
-        # Filter using vectorized mask
-        mask = np.abs(lats.degrees) < 10.0
-        star_data = [(star_names_all[i], star_objs_all[i]) for i in np.where(mask)[0]]
-
-        conjunctions = skyfield_searches.find_conjunctions_with_stars(
-            self.observer,
-            "moon",
-            star_data,
-            self.start_date,
-            self.end_date,
-            threshold_degrees=5.0,
-            precomputed_positions=precomputed_positions,
-        )
-
-        events = []
-        for conj in conjunctions:
-            event_data = {
-                "date": conj["date"].astimezone(utc),
-                "event": "Conjunction",
-                "object1": "Moon",
-                "object2": conj["object2"],
-                "separation_degrees": conj["separation_degrees"],
-                "type": "Moon-Star Conjunction",
-            }
-            event_data["rarity"] = self._get_rarity("Moon-Star Conjunction", event_data)
-            events.append(event_data)
-
-        logger.debug(
-            f"--- calculate_moon_star_conjunctions: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_nasa_comets(self):
-        start_time = time.time()
-        events = []
-        try:
-            comets = cache.get_nasa_comets_data(self.start_date, self.end_date)
-            # Optimization: replace iterrows() with zip() for better performance
-            for name, close_approach_data in zip(
-                comets["name"], comets["close_approach_data"]
-            ):
-                event_data = {
-                    "date": parse_date(
-                        close_approach_data[0]["close_approach_date_full"]  # type: ignore
-                    ).astimezone(utc),
-                    "event": name,
-                    "type": "Comet",
-                }
-                event_data["rarity"] = self._get_rarity("Comet", event_data)
-                events.append(event_data)
-        except Exception as e:
-            logger.error(f"Error calculating NASA comets: {e}")
-        logger.debug(f"--- calculate_nasa_comets: {time.time() - start_time}s")
-        return events
-
-    def calculate_planet_alignments(self):
-        start_time = time.time()
-        events = skyfield_searches.find_planet_alignments(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["type"] = "Planet Alignment"
-            event["rarity"] = self._get_rarity("Planet Alignment", event)
-        logger.debug(f"--- calculate_planet_alignments: {time.time() - start_time}s")
-        return events
-
-    def calculate_golden_blue_hours(self):
-        start_time = time.time()
-        events = skyfield_searches.find_golden_blue_hours(
-            self.observer, self.start_date, self.end_date
-        )
-        # Filter based on settings
-        show_golden = self.event_settings.get("golden_hour")
-        show_blue = self.event_settings.get("blue_hour")
-
-        filtered_events = [
-            e
-            for e in events
-            if (e["type"] == "Golden Hour" and show_golden)
-            or (e["type"] == "Blue Hour" and show_blue)
-        ]
-
-        for event in filtered_events:
-            event["rarity"] = 1  # Common events
-
-        logger.debug(f"--- calculate_golden_blue_hours: {time.time() - start_time}s")
-        return filtered_events
-
-    def calculate_culminations(self):
-        start_time = time.time()
-        events = skyfield_searches.find_culminations(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["rarity"] = 2  # Frequent but useful
-        logger.debug(f"--- calculate_culminations: {time.time() - start_time}s")
-        return events
-
-    def calculate_messier_culminations(self):
-        start_time = time.time()
-        # Check all Messier objects
-        # Optimization: replace iterrows() with zip() for better performance
-        messier_data = list(
-            zip(
-                self.catalogs.MESSIER["Messier"],
-                self.catalogs.MESSIER["skyfield_object"],
-            )
-        )
-
-        events = skyfield_searches.find_object_culminations(
-            self.observer,
-            messier_data,
-            self.start_date,
-            self.end_date,
-        )
-
-        for event in events:
-            event["rarity"] = self._get_rarity("Messier Culmination", event)
-
-        logger.debug(f"--- calculate_messier_culminations: {time.time() - start_time}s")
-        return events
-
-    def calculate_jovian_moon_events(self):
-        start_time = time.time()
-        events = skyfield_searches.find_jovian_moon_events(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Jovian Moon Event", event)
-        logger.debug(f"--- calculate_jovian_moon_events: {time.time() - start_time}s")
-        return events
-
-    def calculate_saturn_ring_crossings(self):
-        start_time = time.time()
-        events = skyfield_searches.find_saturn_ring_crossings(
-            self.start_date, self.end_date
-        )
-        for event in events:
-            # Ensure date is timezone-aware UTC
-            event["date"] = event["date"].astimezone(utc)
-            # These are extremely rare (every 13-15 years), so rarity 5
-            event["rarity"] = 5
-        logger.debug(
-            f"--- calculate_saturn_ring_crossings: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_greatest_elongations(self):
-        start_time = time.time()
-        events = skyfield_searches.find_greatest_elongations(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Greatest Elongation", event)
-        logger.debug(f"--- calculate_greatest_elongations: {time.time() - start_time}s")
-        return events
-
-    def calculate_planetary_dichotomy(self, precomputed_positions=None):
-        start_time = time.time()
-        events = skyfield_searches.find_planetary_dichotomy(
-            self.observer,
-            self.start_date,
-            self.end_date,
-            precomputed_positions=precomputed_positions,
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Planetary Dichotomy", event)
-        logger.debug(f"--- calculate_planetary_dichotomy: {time.time() - start_time}s")
-        return events
-
-    def calculate_seasons(self):
-        start_time = time.time()
-        events = skyfield_searches.find_seasons(self.start_date, self.end_date)
-        for event in events:
-            event["rarity"] = self._get_rarity("Season", event)
-        logger.debug(f"--- calculate_seasons: {time.time() - start_time}s")
-        return events
-
-    def calculate_planet_messier_conjunctions(self, precomputed_positions=None):
-        start_time = time.time()
-        events = skyfield_searches.find_planet_messier_conjunctions(
-            self.observer,
-            self.start_date,
-            self.end_date,
-            precomputed_positions=precomputed_positions,
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Planet-Messier Conjunction", event)
-        logger.debug(
-            f"--- calculate_planet_messier_conjunctions: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_jupiter_grs_transits(self, grs_longitude: Optional[float] = None):
-        start_time = time.time()
-        events = skyfield_searches.find_jupiter_grs_transits(
-            self.observer, self.start_date, self.end_date, grs_longitude=grs_longitude
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Jupiter GRS Transit", event)
-        logger.debug(f"--- calculate_jupiter_grs_transits: {time.time() - start_time}s")
-        return events
-
-    def calculate_planet_star_conjunctions(self, precomputed_positions=None):
-        start_time = time.time()
-        events = skyfield_searches.find_planet_star_conjunctions(
-            self.observer,
-            self.start_date,
-            self.end_date,
-            precomputed_positions=precomputed_positions,
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Planet-Star Conjunction", event)
-        logger.debug(
-            f"--- calculate_planet_star_conjunctions: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_planet_stationary_points(self):
-        start_time = time.time()
-        events = []
-        planets = [
-            "mercury",
-            "venus",
-            "mars barycenter",
-            "jupiter barycenter",
-            "saturn barycenter",
-            "uranus barycenter",
-            "neptune barycenter",
-            "pluto barycenter",
-        ]
-        # Efficiency: Use vectorized approach to find stationary points across all planets
-        # and avoid nested executor submission to prevent potential deadlocks.
-        # Note: Stationary points are searched using a coarse step of 2 days.
-        for p in planets:
-            found_events = skyfield_searches.find_stationary_points(
-                self.observer,
-                p,
-                self.start_date,
-                self.end_date,
-            )
-            for event in found_events:
-                event["rarity"] = self._get_rarity("Planet Stationary Point", event)
-            events.extend(found_events)
-        logger.debug(
-            f"--- calculate_planet_stationary_points: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_planet_solar_conjunctions(self):
-        start_time = time.time()
-        events = skyfield_searches.find_planet_solar_conjunctions(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Planet Solar Conjunction", event)
-        logger.debug(
-            f"--- calculate_planet_solar_conjunctions: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_lunar_features(self):
-        start_time = time.time()
-        events = skyfield_searches.find_lunar_features(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Lunar Feature", event)
-        logger.debug(f"--- calculate_lunar_features: {time.time() - start_time}s")
-        return events
-
-    def calculate_planet_planet_occultations(self):
-        start_time = time.time()
-        events = skyfield_searches.find_planet_planet_occultations(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Planet-Planet Occultation", event)
-        logger.debug(
-            f"--- calculate_planet_planet_occultations: {time.time() - start_time}s"
-        )
-        return events
-
-    def calculate_supermoons(self):
-        start_time = time.time()
-        events = skyfield_searches.find_supermoons(self.start_date, self.end_date)
-        for event in events:
-            # Ensure date is timezone-aware UTC
-            event["date"] = event["date"].astimezone(utc)
-            event["rarity"] = self._get_rarity("Supermoon", event)
-        logger.debug(f"--- calculate_supermoons: {time.time() - start_time}s")
-        return events
-
-    def calculate_mars_closest_approach(self):
-        start_time = time.time()
-        events = skyfield_searches.find_mars_closest_approach(
-            self.start_date, self.end_date, observer=self.observer
-        )
-        for event in events:
-            event["rarity"] = self._get_rarity("Mars Closest Approach", event)
-        logger.debug(f"--- calculate_mars_closest_approach: {time.time() - start_time}s")
-        return events
-
-    def calculate_jovian_mutual_events(self):
-        start_time = time.time()
-        events = skyfield_searches.find_jovian_mutual_events(
-            self.observer, self.start_date, self.end_date
-        )
-        for event in events:
-            # These are rare and high-interest events
-            event["rarity"] = 5
-        logger.debug(f"--- calculate_jovian_mutual_events: {time.time() - start_time}s")
-        return events
