+++ apts/events/__init__.py
@@ -1,0 +1,3 @@
+from .coordinator import AstronomicalEvents
+
+__all__ = ["AstronomicalEvents"]
