+++ tests/unit/test_space_events.py
@@ -22,7 +22,7 @@ class SpaceEventsTest(unittest.TestCase):
             events_to_calculate=[EventType.SPACE_LAUNCHES, EventType.SPACE_EVENTS],
         )
 
-    @patch("apts.events.requests.get")
+    @patch("apts.events.calculations.space.requests.get")
     def test_calculate_space_launches(self, mock_get):
         mock_launch_response = MagicMock()
         mock_launch_response.json.return_value = {
@@ -55,7 +55,7 @@ class SpaceEventsTest(unittest.TestCase):
                 events_df.iloc[0]["date"], datetime(2023, 1, 26, 14, 22, tzinfo=utc)
             )
 
-    @patch("apts.events.requests.get")
+    @patch("apts.events.calculations.space.requests.get")
     def test_calculate_space_events(self, mock_get):
         mock_launch_response = MagicMock()
         mock_launch_response.json.return_value = {"results": []}
@@ -88,7 +88,7 @@ class SpaceEventsTest(unittest.TestCase):
                 events_df.iloc[0]["date"], datetime(2023, 1, 15, 10, 0, tzinfo=utc)
             )
 
-    @patch("apts.events.requests.get")
+    @patch("apts.events.calculations.space.requests.get")
     def test_calculate_space_launches_pl(self, mock_get):
         mock_launch_response = MagicMock()
         mock_launch_response.json.return_value = {
@@ -119,7 +119,7 @@ class SpaceEventsTest(unittest.TestCase):
             # "Space Launch" -> "Start rakiety"
             self.assertEqual(events_df.iloc[0]["type"], "Start rakiety")
 
-    @patch("apts.events.requests.get")
+    @patch("apts.events.calculations.space.requests.get")
     def test_calculate_space_events_pl(self, mock_get):
         mock_launch_response = MagicMock()
         mock_launch_response.json.return_value = {"results": []}
